requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bs4==0.0.1
capstone==4.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
filelock==3.0.12
idna==2.8
importlib-metadata==0.22
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
paramiko==2.6.0
Pillow==6.1.0
pluggy==0.13.0
psutil==5.6.3
py==1.8.0
pycparser==2.19
pyelftools==0.25
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.2
pyserial==3.4
PySocks==1.7.0
requests==2.22.0
ROPGadget==5.8
six==1.12.0
soupsieve==1.9.3
toml==0.10.0
tox==3.14.0
urllib3==1.25.3
virtualenv==16.7.5
zipp==0.6.0
pip3 install --upgrade git+https://github.com/arthaud/python3-pwntools.git
